Darwin_framework_acm/ACM_UI_AUTOMATION_FRAMEWORK/requirements.txt
allure-pytest==2.9.43
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
mysql-connector-python==8.0.28
requests==2.28.1
requests-oauthlib==1.3.0
xmltodict~=0.13.0
pytest-ordering
selenium==4.3.0
webdriver-manager==3.8.3
locator~=0.1.0